requirements.txt
aiogram~=3.17.0
mysql-connector-python~=9.1.0